Renamed Create.js to CreatePoem.js for functionality clarity.
